requirements.txt
pyannote.audio==3.3.2
torch==2.7.0
torchaudio==2.7.0
numpy==2.2.6
speechbrain==1.0.3
tkinterdnd2
ffmpeg-python
librosa>=0.10.1
transformers>=4.36.0
python-dotenv==1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
pydub>=0.25.1
flask>=2.0.1
flask-cors>=4.0.0
python-socketio>=5.10.0
scipy>=1.7.0
datasets>=2.0.0
accelerate>=0.12.0
sentencepiece>=0.1.96
protobuf>=3.20.0
tqdm>=4.64.0
matplotlib>=3.5.0
pandas>=1.4.0
scikit-learn>=1.0.0
webrtcvad>=2.0.10 